pyproject.toml
[project]
name = "sd-defect-demo"
version = "0.1.0"
description = "Controlled diffusion demos for defect data augmentation."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "accelerate>=1.11.0",
    "diffusers>=0.35.0",
    "numpy>=2.2.0",
    "opencv-python>=4.11.0",
    "paramiko>=4.0.0",
    "pillow>=11.0.0",
    "safetensors>=0.6.0",
    "torch>=2.8.0",
    "transformers>=4.56.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = false
